updated config
