+++ express/server.js
@@ -1,1 +1,66 @@
-!function(t,e){for(var r in e)t[r]=e[r]}(exports,function(t){var e={};function r(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return t[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}return r.m=t,r.c=e,r.d=function(t,e,n){r.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:n})},r.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},r.t=function(t,e){if(1&e&&(t=r(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)r.d(n,o,function(e){return t[e]}.bind(null,o));return n},r.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return r.d(e,"a",e),e},r.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},r.p="",r.oe=function(t){process.nextTick((function(){throw t}))},r(r.s=1)}([function(t,e){t.exports=require("json-graphql-parser/v2/index.js")},function(t,e,r){r(2).config();var n=r(3),o=r(4).UpdateHackRPlayBadges,i=r(8).sendMail,a=(r(10).email2Slug,r(11),r(12)),u=r(13),c=n();process.env.SERVER_PORT&&c.set("port",process.env.SERVER_PORT),c.use(u.json());var s="".concat(process.env.NHOST_BACKEND_URL,"/").concat(process.env.NHOST_VERSION,"/").concat(process.env.NHOST_ENDPOINT),l=process.env.SENDGRID_API_KEY;console.log(l),c.get("/",(function(t,e){e.writeHead(200,{"Content-Type":"text/html"}),e.write("<h1>Hello from Express.js!</h1>"),e.end()})),c.post("/badges",(function(t,e){e.writeHead(200,{"Content-Type":"application/json"}),o(s,l).then((function(t){console.log({response:"Successfully updated badges."}),e.end(JSON.stringify(t))}))})),c.post("/mail",(function(t,e){e.writeHead(200,{"Content-Type":"application/json"}),i(l,email2slug("koustov@live.com"));console.log({response:"Successfully sent email."}),e.end(JSON.stringify({result:"done"}))})),t.exports=c,t.exports.handler=a(c)},function(t,e){t.exports=require("dotenv")},function(t,e){t.exports=require("express")},function(t,e,r){function n(t){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function o(){"use strict";/*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */o=function(){return t};var t={},e=Object.prototype,r=e.hasOwnProperty,i=Object.defineProperty||function(t,e,r){t[e]=r.value},a="function"==typeof Symbol?Symbol:{},u=a.iterator||"@@iterator",c=a.asyncIterator||"@@asyncIterator",s=a.toStringTag||"@@toStringTag";function l(t,e,r){return Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{l({},"")}catch(t){l=function(t,e,r){return t[e]=r}}function f(t,e,r,n){var o=e&&e.prototype instanceof d?e:d,a=Object.create(o.prototype),u=new j(n||[]);return i(a,"_invoke",{value:E(t,r,u)}),a}function h(t,e,r){try{return{type:"normal",arg:t.call(e,r)}}catch(t){return{type:"throw",arg:t}}}t.wrap=f;var p={};function d(){}function y(){}function v(){}var m={};l(m,u,(function(){return this}));var g=Object.getPrototypeOf,w=g&&g(g(k([])));w&&w!==e&&r.call(w,u)&&(m=w);var b=v.prototype=d.prototype=Object.create(m);function _(t){["next","throw","return"].forEach((function(e){l(t,e,(function(t){return this._invoke(e,t)}))}))}function x(t,e){var o;i(this,"_invoke",{value:function(i,a){function u(){return new e((function(o,u){!function o(i,a,u,c){var s=h(t[i],t,a);if("throw"!==s.type){var l=s.arg,f=l.value;return f&&"object"==n(f)&&r.call(f,"__await")?e.resolve(f.__await).then((function(t){o("next",t,u,c)}),(function(t){o("throw",t,u,c)})):e.resolve(f).then((function(t){l.value=t,u(l)}),(function(t){return o("throw",t,u,c)}))}c(s.arg)}(i,a,o,u)}))}return o=o?o.then(u,u):u()}})}function E(t,e,r){var n="suspendedStart";return function(o,i){if("executing"===n)throw new Error("Generator is already running");if("completed"===n){if("throw"===o)throw i;return P()}for(r.method=o,r.arg=i;;){var a=r.delegate;if(a){var u=L(a,r);if(u){if(u===p)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if("suspendedStart"===n)throw n="completed",r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);n="executing";var c=h(t,e,r);if("normal"===c.type){if(n=r.done?"completed":"suspendedYield",c.arg===p)continue;return{value:c.arg,done:r.done}}"throw"===c.type&&(n="completed",r.method="throw",r.arg=c.arg)}}}function L(t,e){var r=t.iterator[e.method];if(void 0===r){if(e.delegate=null,"throw"===e.method){if(t.iterator.return&&(e.method="return",e.arg=void 0,L(t,e),"throw"===e.method))return p;e.method="throw",e.arg=new TypeError("The iterator does not provide a 'throw' method")}return p}var n=h(r,t.iterator,e.arg);if("throw"===n.type)return e.method="throw",e.arg=n.arg,e.delegate=null,p;var o=n.arg;return o?o.done?(e[t.resultName]=o.value,e.next=t.nextLoc,"return"!==e.method&&(e.method="next",e.arg=void 0),e.delegate=null,p):o:(e.method="throw",e.arg=new TypeError("iterator result is not an object"),e.delegate=null,p)}function O(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function S(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function j(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(O,this),this.reset(!0)}function k(t){if(t){var e=t[u];if(e)return e.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var n=-1,o=function e(){for(;++n<t.length;)if(r.call(t,n))return e.value=t[n],e.done=!1,e;return e.value=void 0,e.done=!0,e};return o.next=o}}return{next:P}}function P(){return{value:void 0,done:!0}}return y.prototype=v,i(b,"constructor",{value:v,configurable:!0}),i(v,"constructor",{value:y,configurable:!0}),y.displayName=l(v,s,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===y||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,v):(t.__proto__=v,l(t,s,"GeneratorFunction")),t.prototype=Object.create(b),t},t.awrap=function(t){return{__await:t}},_(x.prototype),l(x.prototype,c,(function(){return this})),t.AsyncIterator=x,t.async=function(e,r,n,o,i){void 0===i&&(i=Promise);var a=new x(f(e,r,n,o),i);return t.isGeneratorFunction(r)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},_(b),l(b,s,"Generator"),l(b,u,(function(){return this})),l(b,"toString",(function(){return"[object Generator]"})),t.keys=function(t){var e=Object(t),r=[];for(var n in e)r.push(n);return r.reverse(),function t(){for(;r.length;){var n=r.pop();if(n in e)return t.value=n,t.done=!1,t}return t.done=!0,t}},t.values=k,j.prototype={constructor:j,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=void 0,this.done=!1,this.delegate=null,this.method="next",this.arg=void 0,this.tryEntries.forEach(S),!t)for(var e in this)"t"===e.charAt(0)&&r.call(this,e)&&!isNaN(+e.slice(1))&&(this[e]=void 0)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var e=this;function n(r,n){return a.type="throw",a.arg=t,e.next=r,n&&(e.method="next",e.arg=void 0),!!n}for(var o=this.tryEntries.length-1;o>=0;--o){var i=this.tryEntries[o],a=i.completion;if("root"===i.tryLoc)return n("end");if(i.tryLoc<=this.prev){var u=r.call(i,"catchLoc"),c=r.call(i,"finallyLoc");if(u&&c){if(this.prev<i.catchLoc)return n(i.catchLoc,!0);if(this.prev<i.finallyLoc)return n(i.finallyLoc)}else if(u){if(this.prev<i.catchLoc)return n(i.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<i.finallyLoc)return n(i.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var o=this.tryEntries[n];if(o.tryLoc<=this.prev&&r.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,p):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),p},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.finallyLoc===t)return this.complete(r.completion,r.afterLoc),S(r),p}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.tryLoc===t){var n=r.completion;if("throw"===n.type){var o=n.arg;S(r)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,e,r){return this.delegate={iterator:k(t),resultName:e,nextLoc:r},"next"===this.method&&(this.arg=void 0),p}},t}function i(t,e,r,n,o,i,a){try{var u=t[i](a),c=u.value}catch(t){return void r(t)}u.done?e(c):Promise.resolve(c).then(n,o)}function a(t){return function(){var e=this,r=arguments;return new Promise((function(n,o){var a=t.apply(e,r);function u(t){i(a,n,o,u,c,"next",t)}function c(t){i(a,n,o,u,c,"throw",t)}u(void 0)}))}}var u=r(5).HackRPlayBadges,c=r(6).gsubmit,s=r(7),l=s.GetAllParticipantIdeasMemberIdQuery,f=s.GetUserBadgeQuery,h=s.GetAllAuthorsIdQuery,p=s.GetAllCompletedIdeasIdQuery,d=s.GetAllWinnerUserIdQuery,y=s.InsertBadgeQuery,v=s.GetUserDetailsQuery,m=function(t){var e=[];return t.forEach((function(t){e.indexOf(t.user_id)<0&&e.push(t.user_id)})),e},g=function(t,e){var r=[];return t.forEach((function(t){e.filter((function(e){return e.idea_id===t.idea_id})).length>0&&r.indexOf(t.user_id)<0&&r.push(t.user_id)})),r},w=function(t,e){var r=[];return e.forEach((function(t){"w"===t.position_type.toLowerCase()&&r.push(t.user_id)})),r},b=function(){var t=a(o().mark((function t(e,r,n){var i;return o().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,c(f(e,r),n);case 2:return i=t.sent,console.log(i,i.length,i.length>0),t.abrupt("return",i.length>0);case 5:case"end":return t.stop()}}),t)})));return function(e,r,n){return t.apply(this,arguments)}}(),_=function(){var t=a(o().mark((function t(e,r,n){var i;return o().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,b(e,r,n);case 2:if(t.sent){t.next=11;break}return console.log("Inserting user: ".concat(e)),t.next=6,c(v(e),n);case 6:return i=t.sent,console.log(i),t.abrupt("return",c(y(e,r),n));case 11:console.log("User badge exists: ".concat(e));case 12:case"end":return t.stop()}}),t)})));return function(e,r,n){return t.apply(this,arguments)}}();t.exports.UpdateHackRPlayBadges=function(t,e){var r=[c(l(),t),c(p(),t),c(d(),t),c(h(),t)];return Promise.all(r).then((function(e){var r=e[0],n=e[1],o=e[2],i=e[3],a=r;i.forEach((function(t){a.push({idea_id:t.id,user_id:t.owner})}));var c=m(a),s=g(a,n),l=w(a,o),f=[];return c.forEach((function(e){f.push(_(e,u.participant,t))})),s.forEach((function(e){f.push(_(e,u.gsubmitters,t))})),l.forEach((function(e){f.push(_(e,u.winners,t))})),Promise.all(f).then((function(t){return{winners:l,gsubmitters:s,participants:c}})).catch((function(t){}))})).catch((function(t){}))}},function(t,e){t.exports.HackRPlayBadges={participant:"000d34a9-b675-47f9-80c3-751660970f28",submitters:"000d34a9-b675-47f9-80c3-751660970f29",winners:"000d34a9-b675-47f9-80c3-751660970f30"}},function(t,e,r){function n(t){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function o(){"use strict";/*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */o=function(){return t};var t={},e=Object.prototype,r=e.hasOwnProperty,i=Object.defineProperty||function(t,e,r){t[e]=r.value},a="function"==typeof Symbol?Symbol:{},u=a.iterator||"@@iterator",c=a.asyncIterator||"@@asyncIterator",s=a.toStringTag||"@@toStringTag";function l(t,e,r){return Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{l({},"")}catch(t){l=function(t,e,r){return t[e]=r}}function f(t,e,r,n){var o=e&&e.prototype instanceof d?e:d,a=Object.create(o.prototype),u=new j(n||[]);return i(a,"_invoke",{value:E(t,r,u)}),a}function h(t,e,r){try{return{type:"normal",arg:t.call(e,r)}}catch(t){return{type:"throw",arg:t}}}t.wrap=f;var p={};function d(){}function y(){}function v(){}var m={};l(m,u,(function(){return this}));var g=Object.getPrototypeOf,w=g&&g(g(k([])));w&&w!==e&&r.call(w,u)&&(m=w);var b=v.prototype=d.prototype=Object.create(m);function _(t){["next","throw","return"].forEach((function(e){l(t,e,(function(t){return this._invoke(e,t)}))}))}function x(t,e){var o;i(this,"_invoke",{value:function(i,a){function u(){return new e((function(o,u){!function o(i,a,u,c){var s=h(t[i],t,a);if("throw"!==s.type){var l=s.arg,f=l.value;return f&&"object"==n(f)&&r.call(f,"__await")?e.resolve(f.__await).then((function(t){o("next",t,u,c)}),(function(t){o("throw",t,u,c)})):e.resolve(f).then((function(t){l.value=t,u(l)}),(function(t){return o("throw",t,u,c)}))}c(s.arg)}(i,a,o,u)}))}return o=o?o.then(u,u):u()}})}function E(t,e,r){var n="suspendedStart";return function(o,i){if("executing"===n)throw new Error("Generator is already running");if("completed"===n){if("throw"===o)throw i;return P()}for(r.method=o,r.arg=i;;){var a=r.delegate;if(a){var u=L(a,r);if(u){if(u===p)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if("suspendedStart"===n)throw n="completed",r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);n="executing";var c=h(t,e,r);if("normal"===c.type){if(n=r.done?"completed":"suspendedYield",c.arg===p)continue;return{value:c.arg,done:r.done}}"throw"===c.type&&(n="completed",r.method="throw",r.arg=c.arg)}}}function L(t,e){var r=t.iterator[e.method];if(void 0===r){if(e.delegate=null,"throw"===e.method){if(t.iterator.return&&(e.method="return",e.arg=void 0,L(t,e),"throw"===e.method))return p;e.method="throw",e.arg=new TypeError("The iterator does not provide a 'throw' method")}return p}var n=h(r,t.iterator,e.arg);if("throw"===n.type)return e.method="throw",e.arg=n.arg,e.delegate=null,p;var o=n.arg;return o?o.done?(e[t.resultName]=o.value,e.next=t.nextLoc,"return"!==e.method&&(e.method="next",e.arg=void 0),e.delegate=null,p):o:(e.method="throw",e.arg=new TypeError("iterator result is not an object"),e.delegate=null,p)}function O(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function S(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function j(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(O,this),this.reset(!0)}function k(t){if(t){var e=t[u];if(e)return e.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var n=-1,o=function e(){for(;++n<t.length;)if(r.call(t,n))return e.value=t[n],e.done=!1,e;return e.value=void 0,e.done=!0,e};return o.next=o}}return{next:P}}function P(){return{value:void 0,done:!0}}return y.prototype=v,i(b,"constructor",{value:v,configurable:!0}),i(v,"constructor",{value:y,configurable:!0}),y.displayName=l(v,s,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===y||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,v):(t.__proto__=v,l(t,s,"GeneratorFunction")),t.prototype=Object.create(b),t},t.awrap=function(t){return{__await:t}},_(x.prototype),l(x.prototype,c,(function(){return this})),t.AsyncIterator=x,t.async=function(e,r,n,o,i){void 0===i&&(i=Promise);var a=new x(f(e,r,n,o),i);return t.isGeneratorFunction(r)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},_(b),l(b,s,"Generator"),l(b,u,(function(){return this})),l(b,"toString",(function(){return"[object Generator]"})),t.keys=function(t){var e=Object(t),r=[];for(var n in e)r.push(n);return r.reverse(),function t(){for(;r.length;){var n=r.pop();if(n in e)return t.value=n,t.done=!1,t}return t.done=!0,t}},t.values=k,j.prototype={constructor:j,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=void 0,this.done=!1,this.delegate=null,this.method="next",this.arg=void 0,this.tryEntries.forEach(S),!t)for(var e in this)"t"===e.charAt(0)&&r.call(this,e)&&!isNaN(+e.slice(1))&&(this[e]=void 0)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var e=this;function n(r,n){return a.type="throw",a.arg=t,e.next=r,n&&(e.method="next",e.arg=void 0),!!n}for(var o=this.tryEntries.length-1;o>=0;--o){var i=this.tryEntries[o],a=i.completion;if("root"===i.tryLoc)return n("end");if(i.tryLoc<=this.prev){var u=r.call(i,"catchLoc"),c=r.call(i,"finallyLoc");if(u&&c){if(this.prev<i.catchLoc)return n(i.catchLoc,!0);if(this.prev<i.finallyLoc)return n(i.finallyLoc)}else if(u){if(this.prev<i.catchLoc)return n(i.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<i.finallyLoc)return n(i.finallyLoc)}}}},abrupt:function(t,e){for(var n=this.tryEntries.length-1;n>=0;--n){var o=this.tryEntries[n];if(o.tryLoc<=this.prev&&r.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,p):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),p},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.finallyLoc===t)return this.complete(r.completion,r.afterLoc),S(r),p}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.tryLoc===t){var n=r.completion;if("throw"===n.type){var o=n.arg;S(r)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,e,r){return this.delegate={iterator:k(t),resultName:e,nextLoc:r},"next"===this.method&&(this.arg=void 0),p}},t}function i(t,e,r,n,o,i,a){try{var u=t[i](a),c=u.value}catch(t){return void r(t)}u.done?e(c):Promise.resolve(c).then(n,o)}function a(t){return function(){var e=this,r=arguments;return new Promise((function(n,o){var a=t.apply(e,r);function u(t){i(a,n,o,u,c,"next",t)}function c(t){i(a,n,o,u,c,"throw",t)}u(void 0)}))}}var u="".concat(process.env.NHOST_BACKEND_URL,"/").concat(process.env.NHOST_VERSION,"/").concat(process.env.NHOST_ENDPOINT),c=function(){var t=a(o().mark((function t(e,n,i){var a;return o().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,Promise.resolve().then(r.t.bind(null,0,7));case 2:return a=t.sent,t.abrupt("return",a.submit_multi(e,u));case 4:case"end":return t.stop()}}),t)})));return function(e,r,n){return t.apply(this,arguments)}}(),s=function(){var t=a(o().mark((function t(e,n,i){var a;return o().wrap((function(t){for(;;)switch(t.prev=t.next){case 0:return t.next=2,Promise.resolve().then(r.t.bind(null,0,7));case 2:return a=t.sent,t.abrupt("return",a.submit(e,n,void 0,!0));case 4:case"end":return t.stop()}}),t)})));return function(e,r,n){return t.apply(this,arguments)}}();t.exports.submitMutation=function(t,e){var r=t;return r.object=e,s(r)},t.exports.gsubmit_multi=c,t.exports.gsubmit=s},function(t,e){t.exports.GetAllParticipantIdeasMemberIdQuery=function(){return{display:"Get All Participants",name:"hackathon_ideas_members",function:"hackathon_ideas_members",return:["user_id","idea_id"]}},t.exports.GetUserDetailsQuery=function(t){return{display:"Get User Details",name:"users",function:"users",where:{clause:{operator:"and",conditions:[{field:"id",operator:"eq",value:t}]}},return:["displayName","email"]}},t.exports.GetUserBadgeQuery=function(t,e){return{display:"Get User Badge Map",name:"meta_user_badge_map",function:"meta_user_badge_map",where:{clause:{operator:"and",conditions:[{field:"user_id",operator:"eq",value:t},{field:"badge_id",operator:"eq",value:e}]}},return:["id"]}},t.exports.GetAllAuthorsIdQuery=function(){return{display:"Get All Author",name:"hackathon_ideas",function:"hackathon_ideas",return:["id","owner"]}},t.exports.GetAllCompletedIdeasIdQuery=function(){return{display:"Get All Submissions",name:"hackathon_idea_submission",function:"hackathon_idea_submission",return:["idea_id"]}},t.exports.GetAllWinnerUserIdQuery=function(){return{display:"Get Winners",name:"hackathon_winners",function:"hackathon_winners",return:["user_id","position_type"]}},t.exports.InsertBadgeQuery=function(t,e){return{display:"Get Winners",name:"insert_meta_user_badges_one",function:"insert_meta_user_badges_one",write:!0,object:{badge_id:t,user_id:e},return:["id"]}}},function(t,e,r){var n=r(9);t.exports.sendMail=function(t,e){console.log(">>>>>>>>>>>>>>>>>"),console.log(t),n.setApiKey(t);var r={to:"koustov@gmail.com",from:"ioreactplay@gmail.com",subject:"Sending with SendGrid is Fun",text:"and easy to do anywhere, even with Node.js",html:"<strong>and easy to do anywhere, even with Node.js</strong>",templateId:"d-242034c978464ef7af0508d7c2c1270e",dynamicTemplateData:{for:"Winner",badge_name:"Hack-R-Play Winner",user_email_slug:e}};n.send(r).then((function(){console.log("Email sent")})).catch((function(t){console.error(t)}))}},function(t,e){t.exports=require("@sendgrid/mail")},function(t,e){var r=[{text:".",mask:"__dt__"},{text:"@",mask:"__at__"},{text:"-",mask:"__hy__"}];t.exports.email2Slug=function(t){var e=t;return r.forEach((function(t){e=e.replace(new RegExp("[".concat(t.text,"]"),"gm"),t.mask)})),e}},function(t,e){t.exports=require("os")},function(t,e){t.exports=require("serverless-http")},function(t,e){t.exports=require("body-parser")}]));
+const dotenv = require("dotenv");
+dotenv.config();
+const express = require("express");
+const {
+  UpdateHackRPlayBadges,
+} = require("../models/badges/hack-r-play-2022.js");
+const { sendMail } = require("../services/email/index.js");
+const { email2Slug } = require("../services/util/string.js");
+const os = require("os");
+const serverless = require("serverless-http");
+const bodyParser = require("body-parser");
+
+// Configuration
+var app = express();
+
+// Initialization
+if (process.env.SERVER_PORT) {
+  app.set("port", process.env.SERVER_PORT);
+}
+
+app.use(bodyParser.json());
+
+const BACKEND_URL = `${process.env.NHOST_BACKEND_URL}/${process.env.NHOST_VERSION}/${process.env.NHOST_ENDPOINT}`;
+const SENDGRID_API_KEY = process.env.SENDGRID_API_KEY;
+
+console.log(SENDGRID_API_KEY);
+
+app.get("/", function (req, res) {
+  res.writeHead(200, { "Content-Type": "text/html" });
+  res.write("<h1>Hello from Express.js!</h1>");
+  res.end();
+});
+
+app.post("/badges", function (req, res) {
+  res.writeHead(200, { "Content-Type": "application/json" });
+  UpdateHackRPlayBadges(BACKEND_URL, SENDGRID_API_KEY).then((result) => {
+    var response = {
+      response: "Successfully updated badges.",
+    };
+    console.log(response);
+    res.end(JSON.stringify(result));
+  });
+});
+
+app.post("/mail", function (req, res) {
+  res.writeHead(200, { "Content-Type": "application/json" });
+  sendMail(SENDGRID_API_KEY, email2slug("koustov@live.com"));
+  var response = {
+    response: "Successfully sent email.",
+  };
+  console.log(response);
+  res.end(JSON.stringify({ result: "done" }));
+});
+
+// var server = app.listen(app.get("port"), function () {
+//   var host = server.address().address;
+//   var port = server.address().port;
+
+//   console.log(
+//     `ReactPlay Web Service is running on  https://${os.hostname()}:${port}`
+//   );
+// });
+
+// export default serverless(app);
+module.exports = app;
+module.exports.handler = serverless(app);
