+++ netlify.toml
@@ -1,3 +1,3 @@
 [build]
   command = "npm install && npm run build"
-   functions = "express"
+  functions = "functions"
