+++ package.json
@@ -7,7 +7,7 @@
 	"license": "MIT",
 	"repository": "git@github.com:reactplay/react-play-web-svc.git",
 	"scripts": {
-		"build": "netlify-lambda build . --config webpack.config.js",
+		"build": "netlify-lambda build express --config webpack.config.js",
 		"start": "nodemon server.js"
 	},
 	"dependencies": {
